fer/requirements.txt
torch>=2.1
torchvision>=0.16
timm>=0.9.12
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
opencv-python>=4.8
